+++ R/ggplot_line.R
@@ -104,7 +104,8 @@ theme_line <-
         legend.margin = margin(t = 20, b = 20),
         legend.key = element_rect(fill = "white"),
         legend.key.height = unit(5, "mm"),
-        legend.key.width = unit(5, "mm")
+        legend.key.width = unit(5, "mm"),
+        legend.direction = "vertical" 
       )
     )
   }
@@ -320,8 +321,7 @@ ggplot_line <- function(data,
   }
   else if (isMobile == TRUE) {
     plot <- plot +
-      theme(plot.title.position = "plot") +
-      theme(plot.caption.position = "plot") +
+      theme_mobile_graph() +
       labs(
         title = stringr::str_wrap(title, 40),
         subtitle = stringr::str_wrap(subtitle, 40),
@@ -329,7 +329,7 @@ ggplot_line <- function(data,
         y = stringr::str_wrap(y_title, 30),
         caption = stringr::str_wrap(caption, 50)
       ) +
-      theme(axis.text.x = element_text(hjust = 0.75))
+      theme_mobile_graph()
   }
   
   return(plot)
@@ -584,10 +584,6 @@ ggplot_line_col <-
     }
     else if (isMobile == TRUE) {
       plot <- plot +
-        theme(plot.title.position = "plot") +
-        theme(plot.caption.position = "plot") +
-        theme(legend.position = "bottom") +
-        theme(legend.justification = "left") +
         labs(
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
@@ -596,7 +592,7 @@ ggplot_line_col <-
           caption = stringr::str_wrap(caption, 50)
         )  +
         guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15))) +
-        theme(axis.text.x = element_text(hjust = 0.75)) 
+        theme_mobile_graph()
     }
     
     return(plot)