+++ R/ggplot_point.R
@@ -104,7 +104,8 @@ theme_point <-
         legend.margin = margin(t = 20, b = 20),
         legend.key = element_rect(fill = "white"),
         legend.key.height = unit(5, "mm"),
-        legend.key.width = unit(5, "mm")
+        legend.key.width = unit(5, "mm"),
+        legend.direction = "vertical" 
       )
     )
   }
@@ -320,15 +321,14 @@ ggplot_point <- function(data,
   }
   else if (isMobile == TRUE) {
     plot <- plot +
-      theme(plot.title.position = "plot") +
-      theme(plot.caption.position = "plot") +
       labs(
         title = stringr::str_wrap(title, 40),
         subtitle = stringr::str_wrap(subtitle, 40),
         x = stringr::str_wrap(x_title, 20),
         y = stringr::str_wrap(y_title, 30),
         caption = stringr::str_wrap(caption, 50)
-      )
+      ) + 
+      theme_mobile_graph()
   }
   
   return(plot)
@@ -629,10 +629,6 @@ ggplot_point_col <-
     }
     else if (isMobile == TRUE) {
       plot <- plot +
-        theme(plot.title.position = "plot") +
-        theme(plot.caption.position = "plot") +
-        theme(legend.position = "bottom") +
-        theme(legend.justification = "left") +
         labs(
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
@@ -640,7 +636,8 @@ ggplot_point_col <-
           y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         )  +
-        guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15)))
+        guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15))) +
+        theme_mobile_graph()
     }
     
     return(plot)