+++ R/ggplot_sf.R
@@ -84,7 +84,8 @@ theme_sf <-
         ),
         legend.key = element_rect(fill = "white"),
         legend.key.height = unit(5, "mm"),
-        legend.key.width = unit(5, "mm")
+        legend.key.width = unit(5, "mm"),
+        legend.direction = "vertical" 
       )
     )
   }
@@ -218,13 +219,12 @@ ggplot_sf <- function(data,
   }
   else if (isMobile == TRUE) {
     plot <- plot +
-      theme(plot.title.position = "plot") +
-      theme(plot.caption.position = "plot") +
       labs(
         title = stringr::str_wrap(title, 40),
         subtitle = stringr::str_wrap(subtitle, 40),
         caption = stringr::str_wrap(caption, 50)
-      )
+      ) + 
+      theme_mobile_map()
   }
   
   return(plot)
@@ -478,17 +478,14 @@ ggplot_sf_col <- function(data,
   }
   else if (isMobile == TRUE) {
     plot <- plot +
-      theme(plot.title.position = "plot") +
-      theme(plot.caption.position = "plot") +
-      theme(legend.position = "bottom") +
-      theme(legend.justification = "left") +
       labs(
         title = stringr::str_wrap(title, 40),
         subtitle = stringr::str_wrap(subtitle, 40),
         caption = stringr::str_wrap(caption, 50)
       )  +
       guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15))) +
-      guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15)))  
+      guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15))) +
+      theme_mobile_map()
   }
   
   return(plot)