+++ R/ggplot_vbar.R
@@ -104,7 +104,8 @@ theme_vbar <-
         ),
         legend.margin = margin(t = 20, b = 20),
         legend.key.height = unit(5, "mm"),
-        legend.key.width = unit(5, "mm")
+        legend.key.width = unit(5, "mm"),
+        legend.direction = "vertical" 
       )
     )
   }
@@ -358,8 +359,6 @@ ggplot_vbar <- function(data,
   }
   else if (isMobile == TRUE) {
     plot <- plot +
-      theme(plot.title.position = "plot") +
-      theme(plot.caption.position = "plot") +
       labs(
         title = stringr::str_wrap(title, 40),
         subtitle = stringr::str_wrap(subtitle, 40),
@@ -367,7 +366,7 @@ ggplot_vbar <- function(data,
         y = stringr::str_wrap(y_title, 30),
         caption = stringr::str_wrap(caption, 50)
       ) +
-      theme(axis.text.x = element_text(hjust = 0.75))
+      theme_mobile_graph()
   }
   
   return(plot)
@@ -714,10 +713,6 @@ ggplot_vbar_col <-
     }
     else if (isMobile == TRUE) {
       plot <- plot +
-        theme(plot.title.position = "plot") +
-        theme(plot.caption.position = "plot") +
-        theme(legend.position = "bottom") +
-        theme(legend.justification = "left") +
         labs(
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
@@ -725,17 +720,11 @@ ggplot_vbar_col <-
           y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         ) +
-        guides(fill = guide_legend(
-          ncol = 1,
-          byrow = TRUE,
-          title = stringr::str_wrap(col_title, 15)
-        ), 
-        col = guide_legend(
-          ncol = 1,
-          byrow = TRUE,
-          title = stringr::str_wrap(col_title, 15)
-        )) +
-        theme(axis.text.x = element_text(hjust = 0.75))
+        guides(
+          fill = guide_legend(ncol = 1, title = stringr::str_wrap(col_title, 15)),
+          col = guide_legend(ncol = 1, title = stringr::str_wrap(col_title, 15))
+        ) +
+        theme_mobile_graph()
     }
     
     return(plot)