+++ inst/shiny/template2/data/template2/www/js/mobile.js
@@ -1,0 +1,18 @@
+var isMobileBinding = new Shiny.InputBinding();
+$.extend(isMobileBinding, {
+  find: function(scope) {
+    return $(scope).find(".mobile-element");
+    callback();
+  },
+  getValue: function(el) {
+    return /((iPhone)|(iPod)|(Android)|(BlackBerry))/.test(navigator.userAgent)
+  },
+  setValue: function(el, value) {
+  },
+  subscribe: function(el, callback) {
+  },
+  unsubscribe: function(el) {
+  }
+});
+
+Shiny.inputBindings.register(isMobileBinding);